requirements.txt
flask

flask_cors

flask_material

pymongo

datetime

requests

mysql-connector-python==8.0.17
